% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.character.ueb.R
\name{as.character.ueb}
\alias{as.character.ueb}
\title{as.character.ueb Function}
\usage{
\method{as.character}{ueb}(x)
}
\arguments{
\item{x}{object to be coerced}
}
\value{
attempts to coerce its argument to character type; like as.vector it strips attributes including names. For lists and pairlists (including language objects such as calls) it deparses the elements individually, except that it extracts the first element of length-one character vectors.
}
\description{
Create objects of type "character". Conserve labels
}
\examples{
#create variable
variable <- factor(c(0,0,0,1,1,1,0,"a",1,0, "B"))
Hmisc::label(variable) <- "label de la variable"
# automatic character
variable <- as.character.ueb(variable)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}

Miriam Mota-Foix <mmota.foix@gmail.com>
}
\keyword{class}
\keyword{factor}
\keyword{labels}
\keyword{levels}
\keyword{variable}
