% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_changes.R
\name{desc_changes}
\alias{desc_changes}
\title{desc_changes Function}
\usage{
desc_changes(x, col = "blue")
}
\arguments{
\item{x}{a string}
}
\value{
list_changes list with messages to report.
var_sino <- c("Fum","alcohol")
desc_changes(paste0("Los valores, 0 y 1 se han considerado 'No' y 'Si' respectivamente para las variables: ", paste(var_sino,collapse = ", ") ))
desc_changes("Los valores superiores a 10000 para la variable n_il6_f han sido considerados NA ")
list_changes
#With results = 'asis' rmarkdown chunk
for(i in seq_along(list_changes)){
cat(list_changes[[i]])
}
}
\description{
The function desc_changes is used to save messages of changes applied in database
}
\examples{
desc_changes("")

}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{changes}
\keyword{message}
\keyword{report}
