% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_group.R
\name{desc_group}
\alias{desc_group}
\title{A desc_group Function}
\usage{
desc_group(covariates, group = NULL, data, method = "non-param",
  font_size = 11, width_lev = "8em", byrow = FALSE,
  pval_cut = 0.05, ...)
}
\arguments{
\item{covariates}{a character string with names of variables.}

\item{group}{factor variable. Outcome. Default value is NULL}

\item{data}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables in the model. If they are not found in 'data', the variables are taken from 'environment(formula)'.}

\item{method}{character string indicating the method to test use; possible values are 'param' or 'nonparam'. Default values is 'param'.}

\item{font_size}{A numeric input for table font size}

\item{width_lev}{defines the maximum width of table columns. Default value is 8em}

\item{byrow}{logical or NA. Percentage of categorical variables must be reported by rows (TRUE), by columns (FALSE) or by columns and rows to sum up 1 (NA). Default value is FALSE, which means that percentages are reported by columns (withing groups).}

\item{pval_cut}{cut p.value colored}
}
\description{
DESCRIPCIO DE LA FUNCIO
}
\examples{
 # set.seed(1)
 # data <- df <- data.frame(MUT = factor(c(rep("A", 12),rep("B",13))),
 #                          var = factor(sample(c("Yes", "no"), 25, replace = T)),
 #                          size = rnorm(25),
 #                          id = paste0("a",1:25))
 # data$id <- as.character(data$id)
 # desc_group(group  = "MUT",covariates = c("var","size","id"), data = data)
}
\keyword{ci}
\keyword{descriptive}
\keyword{exploratory}
\keyword{qualitative}
\keyword{summary}
