% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_group_numeric.R
\name{desc_group_numeric}
\alias{desc_group_numeric}
\title{A desc_group_numeric Function}
\usage{
desc_group_numeric(
  data,
  covariates,
  y = NULL,
  frml = NULL,
  method = "non-param",
  method.cor = "pearson",
  caption = NULL,
  font_size = 13,
  width_lev = "9em",
  col.background = "#993489",
  show.pval = TRUE,
  show.all = TRUE,
  show.n = TRUE,
  corplot = FALSE,
  nround = 1,
  cor_cut = 0.7,
  pval_cut = 0.05
)
}
\arguments{
\item{data}{data frame que contiene las variables a analizar.}

\item{covariates}{variables explicativas a evaluar. Puede indicarse mediante selección tidyverse.}

\item{y}{variable o variables respuesta. Puede indicarse mediante selección tidyverse.}

\item{frml}{fórmula del tipo y ~ x1 + x2 + ... . Si se especifica, tiene prioridad sobre los argumentos \code{covariates} y \code{y}.}

\item{method}{método para la comparación de variables numéricas según grupos. Por defecto \code{"non-param"}.}

\item{method.cor}{método de correlación utilizado para variables numéricas. Por defecto \code{"pearson"}. Falta incluir valores que acepta. Spearman no funciona}

\item{caption}{título de la tabla. Si es NULL se genera automáticamente.}

\item{font_size}{tamaño de letra de la tabla html. Por defecto 13.}

\item{width_lev}{ancho de la columna de niveles. Por defecto \code{"9em"}.}

\item{col.background}{color de fondo de la cabecera de la tabla html.}

\item{show.pval}{TRUE o FALSE indicando si se muestran los p-valores. Por defecto TRUE.}

\item{show.all}{TRUE o FALSE indicando si se muestran las filas resumen "ALL". Por defecto TRUE.}

\item{show.n}{TRUE o FALSE indicando si se muestra el tamaño muestral. Por defecto TRUE.}

\item{corplot}{TRUE o FALSE indicando si se genera gráfico de correlación. Por defecto FALSE.}

\item{nround}{número de decimales para mostrar los resultados. Por defecto 1.}

\item{cor_cut}{punto de corte para seleccionar variables numéricas según el coeficiente de correlación. Por defecto 0.7.}

\item{pval_cut}{punto de corte para seleccionar variables categóricas según el p-valor. Por defecto 0.05.}
}
\value{
Una lista con:
\itemize{
  \item \code{outcome}: variable o variables respuesta analizadas.
  \item \code{covariates}: variables explicativas evaluadas.
  \item \code{df_all}: tabla resumen en formato data.frame.
  \item \code{selVar}: variables seleccionadas según los criterios definidos.
  \item \code{results}: tabla formateada en html mediante kableExtra.
}
}
\description{
Genera una tabla resumen para covariables numéricas y factores respecto a una o varias variables respuesta numéricas.
Es la funcion equivalente a desc_group pero cuando tu variable outcome o de interés es numérica.
Para covariables numéricas calcula correlaciones y para covariables categóricas realiza comparaciones entre grupos,
mostrando únicamente las variables que cumplen los criterios de selección definidos.
}
\examples{
data(mtcars)
mtcars$cyl <- factor(mtcars$cyl)
mtcars$am <- factor(mtcars$am)

# Una variable respuesta
desc_group_numeric(
  data = mtcars,
  covariates = c(mpg, wt, cyl),
  y = disp
)

# Utilizando fórmula
desc_group_numeric(
  data = mtcars,
  frml = disp ~ mpg + wt + cyl
)

# Varias variables respuesta
desc_group_numeric(
  data = mtcars,
  covariates = c(mpg, wt, cyl),
  y = c(disp, drat)
)

# Modificando puntos de corte
desc_group_numeric(
  data = mtcars,
  covariates = c(mpg, wt, cyl),
  y = disp,
  cor_cut = 0.5,
  pval_cut = 0.10
)

}
\author{
Àlex Martí Barrera, Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{correlations}
\keyword{descriptives}
\keyword{summary}
\keyword{tables}
