% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.test.R
\name{diag.test}
\alias{diag.test}
\title{A diag.test Function}
\usage{
diag.test(pred, y, tag.healthy = levels(y)[1], nround = 2)
}
\arguments{
\item{pred}{variable factor outcome predit.}

\item{y}{variable factor outcome original}

\item{tag.healthy}{the value codifying healthy individuals in the status variable. Por defecto nivel de referencia levels(dat[,group])[1]}

\item{nround}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed}
}
\value{
variable$table:  tabla 2x2 de la predicción y referencia.

variable$positive.class:  factor level that corresponds to a "positive" result (if that makes sense for your data)

summary: taula detallada amb totes les mesures de classificació per a cada una de les variables
}
\description{
mesures de clasificació
}
\examples{
# data(aSAH)
# res_roc <- doROC (frml = outcome ~ s100b, title = 'prova1',
# cex.main = 0.6, dat = aSAH, modGLM = TRUE)
# diag.test(y = aSAH$outcome, pred = res_roc$dat$outcome.predict, tag.healthy = "Good"  )

# df_pred <- data.frame(p1 = res_roc$dat$outcome.predict,
# p2 = factor(ifelse(aSAH$s100b >= 0.3, levels(aSAH$outcome)[2], levels(aSAH$outcome)[1]  )))
# diag.test(y = aSAH$outcome, pred = df_pred)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{glm}
\keyword{roc}
\keyword{test}
