% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_names.R
\name{extract_names}
\alias{extract_names}
\title{A extract_names Function}
\usage{
extract_names(varlev, var, del.intercept = T)
}
\arguments{
\item{varlev}{character vector. Nombre de la variable (var) y los niveles de los factores. Habitualmente, la salida de los coeficientes de un modelo.}

\item{var}{character vector. Original name of variable}

\item{del.intercept}{TRUE or FALSE. Remove "(Intercept)"}
}
\description{
Genera resumen para los coeficientes de laso distintos de 0 ( o todos)
}
\examples{
ch2clean <- c("otras_drogasSí",   "svp_ostomiaSí"  ,     "destino_al_altaCSS" , "localesSí"  ,
"fractura_luxacionSí", "edad"  ,   "minutos_cirugia"  ,   "fijacion_externaSi" )
var_orig <- c("variable1", "variable2" ,"otras_drogas" ,   "sexo", "hipertension",  "svp_ostomia"  ,     "destino_al_alta" ,  "locales"  ,
 "fractura_luxacion", "edad"    ,          "minutos_cirugia" ,  "fijacion_externa" )
extract_names (ch2clean, var_orig)
}
\keyword{coefficients}
\keyword{lasso}
\keyword{models}
