% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_ueb.R
\name{factor_ueb}
\alias{factor_ueb}
\title{factor_ueb Function}
\usage{
factor_ueb(x, levels, labels)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct values.}

\item{levels}{an optional vector of the values (as character strings) that x might have taken. The default is the unique set of values taken by as.character(x), sorted into increasing order of x. Note that this set can be specified as smaller than sort(unique(x))}

\item{labels}{either an optional character vector of (unique) labels for the levels (in the same order as levels after removing those in exclude), or a character string of length 1.}
}
\value{
factor returns an object of class "factor" which has a set of integer codes the length of x with a "levels" attribute of mode character and unique (!anyDuplicated(.)) entries. If argument ordered is true (or ordered() is used) the result has class c("ordered", "factor"). Undocumentedly for a long time, factor(x) loses all attributes(x) but "names", and resets "levels" and "class".
}
\description{
The function factor is used to encode a vector as a factor (the terms ‘category’ and ‘enumerated type’ are also used for factors).
}
\examples{
variable <- c(0,0,0,1,1,1,0,"a",1,0)
factor_ueb(variable, levels = c(0,1), labels = c("No", "Si"))

variable <- c(0,0,0,1,1,1,0,10,1,0)
factor_ueb(variable, levels = c(0,1), labels = c("No", "Si"))

variable <- c(0,0,0,1,1,1,0,1,1,0)
factor_ueb(variable, levels = c(0,1), labels = c("No", "Si"))
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{class}
\keyword{factor}
\keyword{labels}
\keyword{levels}
\keyword{variable}
