% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{A freq_table Function}
\usage{
freq_table(
  var.cat,
  y,
  data,
  margin = NULL,
  nround = 2,
  restab = TRUE,
  title = "titles for the table",
  xtab.type = "html"
)
}
\arguments{
\item{var.cat}{vector con el nombre de las variables categoricas a resumir}

\item{y}{nombre de la variable factor principal.}

\item{data}{data frame que contiene las variables a analizar}

\item{margin}{index, or vector of indices to generate margin for}

\item{nround}{integer indicating the number of decimal places (round).}

\item{restab}{TRUE o FALSE, para obtener tabla en formato .tex or .html}

\item{title}{titles for the table}

\item{xtab.type}{A character string. Possible values are latex, html, markdown, pandoc, and rst; this will be automatically determined if the function is called within knitr; it can also be set in the global option knitr.table.format. If format is a function, it must return a character string.}
}
\description{
Tabla de frecuencias relativas, por fila columnas o global.
}
\examples{
# freq_table(var.cat = c('cyl','vs','gear', 'carb'), y = 'am', data = mtc_bis, restab = TRUE)
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{compare}
\keyword{freq}
\keyword{qualitative}
\keyword{tables}
