% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lab_nam.R
\name{get_lab_nam}
\alias{get_lab_nam}
\title{Obtener etiquetas o nombres de variables en un data frame}
\usage{
get_lab_nam(data, variable_names)
}
\arguments{
\item{data}{Un `data.frame` que contiene las variables.}

\item{variable_names}{Un vector de caracteres con los nombres de las variables que se desean consultar.}
}
\value{
Un vector de caracteres con las etiquetas de las variables si están disponibles,
o los nombres de las variables si no tienen etiqueta asignada.
}
\description{
Esta función devuelve las etiquetas (`label`) de una serie de variables en un `data.frame`
si existen. Si alguna variable no tiene una etiqueta, se devuelve el nombre de la variable.
}
\examples{
library(Hmisc)
data <- data.frame(id = 1:3, sexe = c("M", "F", "F"))
label(data$id) <- "Identificador del Paciente"
get_lab_nam(data, c("id", "sexe"))
# Resultado: "Identificador del Paciente" "sexe"

}
