% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color.R
\name{gg_color}
\alias{gg_color}
\title{A gg_color Function}
\usage{
gg_color(n)
}
\arguments{
\item{n}{número total de colores a generar}
}
\description{
Genera el código de colores que usa la libreria ggplot
}
\examples{
gg_color(3)
}
\keyword{color}
\keyword{descriptive}
\keyword{gg}
\keyword{ggplot}
\keyword{plots}
