% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_ueb.R
\name{kable_ueb}
\alias{kable_ueb}
\title{A kable_ueb Function}
\usage{
kable_ueb(tab, caption = NULL, font_size = 13, full_width = FALSE,
  ...)
}
\arguments{
\item{tab}{table to kable}

\item{caption}{Character vector containing the table's caption or title. Set to NULL to suppress the caption. Default value is NULL.}

\item{font_size}{A numeric input for table font size}

\item{full_width}{A TRUE or FALSE variable controlling whether the HTML table should have 100% width. Since HTML and pdf have different flavors on the preferable format for full_width. If not specified, a HTML table will have full width by default but this option will be set to FALSE for a LaTeX table}
}
\description{
DESCRIPCIO DE LA FUNCIO
}
\examples{
aa <- factor(rep(c("A","B","C"), 10))
kable_ueb(table(aa))
kable_ueb(table(aa,aa))
}
\keyword{html}
\keyword{kable}
\keyword{ueb}
