% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_ueb.R
\name{list_files_ueb}
\alias{list_files_ueb}
\title{Llista arxius de diferents carpetes de dades.}
\usage{
list_files_ueb(
  home_folder,
  recursive = TRUE,
  pattern = c("xlsx", "xls", "csv", "sav")
)
}
\arguments{
\item{home_folder}{Objecte R amb la ruta a les carpetes de dades.}

\item{recursive}{Lògic (TRUE per defecte). Llista els arxius de forma recursiva.}

\item{pattern}{Carrega un format d'extensió específic entre 'xlsx', 'xls' i 'csv'.}
}
\value{
Una llista d'arxius estructurada per 'carpetes' anomenades.
}
\description{
Llista arxius de diferents carpetes de dades donat un 'home_folder'.
}
\author{
Miquel Vázquez-Santiago \email{miquel.vazquez@vhir.org}
Biomedical Data Intelligence Unit (BIDU)
Vall d'Hebron Research Institute (VHIR)

Modificaciones y mantenimiento:
Biomedical Data Intelligence Unit (BIDU) 
Vall d'Hebron Research Institute (VHIR) | 
Vall d'Hebron Barcelona Hospital Campus.
}
