% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{A makeTransparent Function}
\usage{
makeTransparent(..., alpha = 0.5)
}
\arguments{
\item{alpha}{nivel de translucido}

\item{color}{nombre del color en formato R}
}
\description{
Creación de colores transparentes
ROC(representación gráfica de la sensibilidad en frente a la especificidad).
}
\examples{
makeTransparent('mediumorchid4')
}
\keyword{color}
\keyword{transparent}
