% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_ueb.R
\name{message_ueb}
\alias{message_ueb}
\title{message_ueb Function}
\usage{
message_ueb(x, col = "blue", ls = FALSE)
}
\arguments{
\item{x}{a string}

\item{col}{color}

\item{ls}{si se trata del elemento de una lista}
}
\description{
The function desc_changes is used to save messages of changes applied in database
}
\examples{
message_ueb("se ha realizado un cambio en el nombre de la variable", col = "purple")


}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{changes}
\keyword{message}
\keyword{report}
