% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_bis-data.R
\docType{data}
\name{mtc_bis}
\alias{mtc_bis}
\title{Motor Trend Car Road Tests. (reclass)}
\format{
A data frame with 32 observations on 11 variables.
}
\description{
The data was extracted from the 1974 Motor Trend US magazine,
and comprises fuel consumption and 10 aspects of automobile design
and performance for 32 automobiles (1973–74 models).
}
\details{
\code{\link{mtcars}}
}
\keyword{datasets}
