% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAperc.R
\name{NAperc}
\alias{NAperc}
\title{A NAperc Function}
\usage{
NAperc(df, maxNA = 80)
}
\arguments{
\item{df}{data frame que contiene las variables a evaluar}

\item{maxNA}{Porcentage de NA permitido}
}
\description{
Percentatge de dades faltants per variable (na$perc) i variables amb més de 'x' percentatge(na$var).
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{missing}
\keyword{na}
\keyword{percentage}
