% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_roc.R
\name{plot_multi_roc}
\alias{plot_multi_roc}
\title{Graficar múltiples curvas ROC de modelos}
\usage{
plot_multi_roc(
  modelos,
  datos,
  outcome,
  colores = NULL,
  text.xlim = 0.15,
  text.ylim = 0.7
)
}
\arguments{
\item{modelos}{Una lista de modelos de predicción (por ejemplo, modelos ajustados de regresión logística,
árboles de decisión (pendiente de hacer pruebas), etc.). Cada elemento de la lista debe tener un nombre que será utilizado en la leyenda del gráfico.}

\item{datos}{Un data frame que contiene los datos que se usarán para hacer las predicciones. El data frame debe incluir
una columna con las etiquetas de la variable de respuesta (outcome).}

\item{outcome}{El nombre de la columna en `datos` que representa la variable dependiente (la variable de clase binaria: caso/control).}

\item{colores}{Un vector de caracteres (opcional) con los colores a utilizar para graficar las curvas ROC de cada modelo. Si no se especifica,
se generarán colores predeterminados automáticamente.}
}
\value{
No retorna nada. La función genera un gráfico de las curvas ROC en el dispositivo gráfico de R.
}
\description{
Esta función genera un gráfico que muestra las curvas ROC (Receiver Operating Characteristic)
para varios modelos, comparándolos en un solo gráfico. Además, muestra los valores AUC (Área Bajo la Curva)
para cada modelo en la leyenda y proporciona una indicación de los niveles de control y caso en el conjunto de datos.
}
\details{
1. Si no se proporcionan colores, la función asigna un conjunto de colores predeterminados utilizando la función `gg_color()`.
2. Para cada modelo, se predicen las probabilidades utilizando `predict()`, y luego se calcula la curva ROC con la función `roc()` de la librería `pROC`.
3. Las curvas ROC se grafican en el mismo espacio de coordenadas y se añaden sucesivamente a la gráfica.
4. La leyenda muestra los nombres de los modelos junto con sus valores AUC.
5. Se incluye texto adicional que indica los niveles de "control" y "caso" en el gráfico.
}
\examples{
# Cargar librerías necesarias
# library(pROC)

# Crear algunos modelos (por ejemplo, regresión logística y un árbol de decisión)
# modelo1 <- glm(outcome ~ ., data = datos_train, family = binomial)
# modelo2 <- rpart(outcome ~ ., data = datos_train, method = "class")

# Graficar múltiples curvas ROC
# plot_multiple_roc(
 #  modelos = list(Modelo_Logistico = modelo1, Modelo_Arbol = modelo2),
 #  datos = datos_test,
 #  outcome = "outcome"
 # )

}
