% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTexCG.R
\name{postTexCG}
\alias{postTexCG}
\title{A postTexCG Function}
\usage{
postTexCG(nameFile)
}
\arguments{
\item{nameFile}{Character vector of length 1 containing the name of file.}
}
\value{
Reescribe el fichero eliminando "2*"
}
\description{
Corrige problema con funcion export2latex
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{compareGroups}
\keyword{error}
\keyword{export2latex}
\keyword{tex}

