% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickSurv.R
\name{quickSurv}
\alias{quickSurv}
\title{quickSurv Function}
\usage{
#' @usage \\method{names}{mtc_bis}(x) <- value
}
\arguments{
\item{time}{for right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval.}

\item{event}{the status indicator. Factor with two levels}

\item{tag.event}{A character value. Level of event}

\item{covariate}{variables a evaluar}

\item{data}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula.}

\item{do.plot}{A logical value indicating if show plot. Default value is TRUE.}

\item{pval.plot}{A logical value indicating if show pvalue at plot. Default value is TRUE.}

\item{title.plot}{a title for the plot.}

\item{summary.km}{type of table to produce for Kaplan-Meier summary. Possible values for type are "ht", "r" or "no". Default value is "ht".}

\item{do.cox}{A logical value indicating if show cox regression. Default value is TRUE.}

\item{summary.cox}{type of table to produce for Kaplan-Meier summary. Possible values for type are "ht", "r" or "no". Default value is "ht".}

\item{title.cox}{a title for the Cox regression.}

\item{pval.cut}{Pvalor que s'utlitza per mostrar el Cox. Per defecte 0.05}
}
\value{
results:

coeff:

plot
}
\description{
Realiza supervivencia Kaplan Meier, cox para variables cuali y cuanti
}
\examples{

# set.seed(81)
# dat <- data.frame(tm = c(rnorm(50,12*4,19), rnorm(50,12*2,12)),
# ev = factor(sample(c("Vivo", "Muerto"), 100, replace = TRUE)) ,
#                   iq = factor(c(rep("A",50), rep("B",50))), age = rnorm(100,22,6))
#
# quickSurv(time = tm,event = ev, covariate = age, tag.event = "Muerto",
#          data = dat, do.cox = TRUE,summary.cox = "ht")


# quickSurv(time = tm,event = ev, covariate = iq, tag.event = "Muerto",
#           data = dat, surv.median.line = "hv",
#           cox = TRUE, summary.cox = "ht" )

# quickSurv(time = tm,event = ev, covariate = iq, tag.event = "Muerto",
#           data = dat, surv.median.line = "hv", cumevents = TRUE,
#          cox = TRUE, summary.cox = "ht")
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{correlation}
\keyword{pearson}
\keyword{plotcor}
\keyword{quickCor}
\keyword{sperman}
