% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resCG.R
\name{resCG}
\alias{resCG}
\title{resCG Function}
\usage{
resCG(restab, dat, y, xtab = FALSE, col = TRUE, title = NULL,
  lbl = NULL)
}
\arguments{
\item{restab}{an object of class 'createTable'.}

\item{dat}{a data frame containing the variables in the model.}

\item{y}{a vector variable that distinguishes the groups. It must be either a numeric, character, factor or NULL. Default value is NULL which means that descriptives for whole sample are calculated and no test is performed}

\item{xtab}{A logical value indicating whether the output is a xtable}

\item{col}{A logical value indicating the xtable color.}

\item{title}{Character vector containing the table's caption or title. Default value is NULL.}

\item{lbl}{Character vector of length 1 containing the LaTeX label. Default value is NULL.}
}
\value{
summary table.
}
\description{
Summary of restab$descr
}
\examples{
res <- compareGroups(am ~., data = mtc_bis, method = NA)
restab <- createTable(res)
resCG(restab, mtc_bis, y = "am", xtab = FALSE)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{comparegroups}
\keyword{summary}
\keyword{tests}

