% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepLR.R
\name{stepLR}
\alias{stepLR}
\title{A stepLR Function}
\usage{
stepLR(VR, varExpl, data, var2mod = NA, trace = TRUE, thrPval = 0.1)
}
\arguments{
\item{VR}{Nombre de la variable respuesta. Debe ser una variable categorica dicotomica}

\item{varExpl}{vector con los nombre de las variables explicativas que se quieren tener en cuenta para construir el modelo.}

\item{data}{data.frame con los datos a trabajar}

\item{var2mod}{Nombre de las variables explicativas a incluir en el modelo (sean o no significativas). En el caso de no indicar ninguna, se partira del modelo nulo}

\item{trace}{TRUE si se quiere mostrar la construcción del modelo paso a paso}

\item{thrPval}{valor a partir del cual se considerara que una variable es significativa (p.valor resultante de comparar dos modelos). Por defecto 0.1}
}
\value{
modfin: exporta dos modelos, el final teniendo en cuenta solo aquellos individuos que tienen todos los valores para todas las variables y un segundo modelo que tiene en cuenta todos aquellos individuos que tienen valores para las variables finales.
}
\description{
Construye el mejor modelo de regresion logistica basandose en el AIC y cuyo pvalor
al comparar los modelos mediante LRT sea inferior a 0.1,  partiendo de 'x' variables explicativas.
IMPORTANTE: para la construcción de todos los modelos se tendran en cuenta solo aquellos pacientes que tengan valores para todas las variables indicadas excepto para el final que tendra
en cuenta las variables finales.
}
\examples{
# modfin <- stepLR(VR = 'vs',varExpl = c('hp', 'am', 'carb'), data = mtc_bis, trace = TRUE )
# summary(modfin[[1]])
# summary(modfin[[2]])
}
\author{
Miriam Mota \email{mmota.foix@gmail.com}
}
\keyword{aic}
\keyword{forward}
\keyword{glm}
\keyword{logisitica}
\keyword{regresion}
\keyword{step}
\keyword{stepwise}
