% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCG.R
\name{summaryCG}
\alias{summaryCG}
\title{summaryCG Function}
\usage{
summaryCG(res, restab, dat, y, xtab = FALSE, col = TRUE, title = NULL,
  lbl = NULL, met.adj = "fdr")
}
\arguments{
\item{res}{an object of class 'compareGroups'}

\item{restab}{an object of class 'createTable' (of param "res").}

\item{dat}{a data frame containing the variables in the model.}

\item{y}{a vector variable that distinguishes the groups. It must be either a numeric, character, factor or NULL. Default value is NULL which means that descriptives for whole sample are calculated and no test is performed}

\item{xtab}{A logical value indicating whether the output is a xtable}

\item{col}{A logical value indicating the xtable color.}

\item{title}{Character vector containing the table's caption or title. Default value is NULL.}

\item{lbl}{Character vector of length 1 containing the LaTeX label. Default value is NULL.}

\item{met.adj}{correction method. Can be abbreviated. Default value is "fdr"}
}
\value{
summary table.
}
\description{
Summary of restab$descr
}
\details{
The adjustment methods include the Bonferroni correction ("bonferroni") in which the 
p-values are multiplied by the number of comparisons. Less conservative corrections
 are also included by Holm (1979) ("holm"), Hochberg (1988) ("hochberg"), Hommel 
 (1988) ("hommel"), Benjamini & Hochberg (1995) ("BH" or its alias "fdr"), 
 and Benjamini & Yekutieli (2001) ("BY"), respectively.
}
\examples{
res <- compareGroups(am ~., data = mtc_bis, method = NA)
restab <- createTable(res)
summaryCG(res, restab,  dat = mtc_bis, y = "am", xtab = FALSE)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{comparegroups}
\keyword{summary}
\keyword{tests}

