% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsurv_df.R
\name{sumsurv_df}
\alias{sumsurv_df}
\title{A sumsurv_df Function}
\usage{
sumsurv_df(sumsurv)
}
\arguments{
\item{sumsurv}{summary.survfit object}
}
\description{
data frame of summary.survfit object
}
\keyword{summary}
\keyword{survfit}
\keyword{survival}
