% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_model_segmented.R
\name{tab_model_segmented}
\alias{tab_model_segmented}
\title{Crear una tabla sjPlot::tab_model limpia para un modelo segmentado}
\usage{
tab_model_segmented(mod_segm, show.df = F)
}
\arguments{
\item{mod_segm}{Un objeto de clase `segmented`, creado con la función `segmented()`.}
}
\value{
Una tabla HTML creada con `sjPlot::tab_model()` que muestra los coeficientes del modelo,
incluyendo los puntos de corte `psi` con sus estimaciones correctas.
}
\description{
Esta función toma un modelo segmentado (`segmented::segmented`) y prepara una tabla con los coeficientes,
reemplazando los valores `Estimate` de los puntos de corte (`psi`) por los valores estimados reales,
ya que sjPlot::tab_model no los muestra correctamente por defecto.
}
\examples{
#modelo_segmentado <- segmented(lm(y ~ x, data = datos), seg.Z = ~x, npsi = 1)
#tab_model_segmented(modelo_segmentado)

}
