% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_mod.R
\name{tabOR_lr}
\alias{tabOR_lr}
\title{A desc_mod Function}
\usage{
tabOR_lr(...)
}
\arguments{
\item{mod}{a fitted object of class inheriting from "glm", "clogit","cox" or "lm".}

\item{xtab}{TRUE o FALSE, para obtener tabla en formato .tex}

\item{title}{if xtab = T, Character vector containing the table's caption or title.}

\item{label}{Character vector of length 1 containing the LaTeX label. Default value is NULL.}

\item{show.intcp}{TRUE o FALSE, indica si se muestra o no el intercept del modelo. En ambos casos el modelo se ha calcula con intercept. Default value is "FALSE".}

\item{show.n}{TRUE o FALSE muestra el total de individuos usados para el ajuste del modelo. Default value is "TRUE".}

\item{show.aov.pval}{TRUE o FALSE muestra el p-valor del modelo global. Default value is "TRUE".}

\item{show.pretty}{TRUE o FALSE muestra las 'labels' de las variables. Solo funciona para lm y glm . Default value is "FALSE".}

\item{group_rw}{TRUE o FALSE  agrupa las filas por variables. Default value is "FALSE".}

\item{row.names}{TRUE or FALSE. Show or not rownames}

\item{digits}{numeric vector specifying the number of digits to display}
}
\description{
Resume y formatea modelos logísticos, lineales, de supervivencia, coxph, binomial negativa y clogit, tanto univariados como multivariados.
}
\keyword{OR}
\keyword{logistica}
\keyword{regresion}
