% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_to_num.R
\name{var_to_num}
\alias{var_to_num}
\title{var_to_num Function}
\usage{
var_to_num(x)
}
\arguments{
\item{x}{a character variable with numbers}
}
\value{
a numeric variable and a warning with delete values.
 edad <- c("6", "6,7", "10 ", "12.5", "3   ", "4,", "no se" )
 Hmisc::label(edad) <- "Edad de los pacientes"
 edad_num <- var_to_num(edad)
 edad_num
 Hmisc::label(edad_num)
}
\description{
The function var_to_num is used to convert a character variable to numeric variable. Change ',' by '.'. Delete spaces. Return a numeric variable and a warning with the delete values.
}
\examples{

}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{change}
\keyword{character}
\keyword{class}
\keyword{numeric}
\keyword{warnings}
