% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_data.R
\name{desc_data}
\alias{desc_data}
\title{A desc_data Function}
\usage{
desc_data(data, format = "html", maxlev = 7, maxNA = 80, size = 13,
  caption = NULL, remove_cols = TRUE)
}
\arguments{
\item{data}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables in the model. If they are not found in 'data', the variables are taken from 'environment(formula)'.}

\item{format}{a character string; possible values are latex, html, markdown, pandoc, and rst; this will be automatically determined if the function is called within knitr; it can also be set in the global option knitr.table.format; if format is a function, it must return a character string}

\item{maxlev}{an integer indicating the maximum number of levels of variable. Default value is 7.}

\item{maxNA}{an integer indicating the maximum number of missing data.}

\item{size}{A numeric input for table font size}

\item{caption}{Character vector containing the table's caption or title. Default value is a summary.}

\item{remove_cols}{logical value. Removes all columns from a data that are composed entirely of NA values.}
}
\description{
DESCRIPCIO DE LA FUNCIO
}
\examples{
# desc_data(data = iris, format = "html")
# desc_data(airquality, format = "html", maxNA = 20)
}
\keyword{clean}
\keyword{data}
\keyword{depurate}
\keyword{read}
\keyword{summary}
