% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_group.R
\name{descGroup}
\alias{descGroup}
\title{A desc_group Function}
\usage{
descGroup(...)
}
\arguments{
\item{frml}{Right side of ~ must have the terms in an additive way, and left side of ~ must contain the name of the grouping variable or can be left in blank (in this latter case descriptives for whole sample are calculated and no test is performed).}

\item{covariates}{a character string with names of variables.}

\item{group}{factor variable. Outcome. Default value is NULL}

\item{data}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables in the model. If they are not found in 'data', the variables are taken from 'environment(formula)'.}

\item{method}{character string indicating the method to test use; possible values are 'param' or 'nonparam'. Default values is 'param'.}

\item{font_size}{A numeric input for table font size}

\item{byrow}{logical or NA. Percentage of categorical variables must be reported by rows (TRUE), by columns (FALSE) or by columns and rows to sum up 1 (NA). Default value is FALSE, which means that percentages are reported by columns (withing groups).}

\item{width_lev}{defines the maximum width of table columns. Default value is 8em}

\item{pval_cut}{cut p.value colored}

\item{show.pval}{logical indicating whether p-value of overall groups significance ('p.overall' column) is displayed or not. Default value is TRUE.}

\item{show.pval.adj}{logical indicating whether adjust p-value of overall groups significance ('p.overall' column) is displayed or not. Default value is FALSE.}

\item{caption}{Character vector containing the table's caption or title.}
}
\description{
DESCRIPCIO DE LA FUNCIO
}
\examples{
 # set.seed(1)
  dat <- df <- data.frame(MUT = factor(c(rep("A", 12),rep("B",13))),
                           var = factor(sample(c("Yes", "no"), 25, replace = TRUE)),
                           size = rnorm(25),
                           id = paste0("a",1:25))
  dat$id <- as.character(dat$id)
  res <- desc_group(group  = "MUT",covariates = names(dat), data = dat)
  res$res
}
\keyword{ci}
\keyword{descriptive}
\keyword{exploratory}
\keyword{qualitative}
\keyword{summary}
