% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCor.R
\name{quickCor}
\alias{quickCor}
\title{quickCor Function}
\usage{
quickCor(x, y, dat, nround = 3, xtab = TRUE, pos = "bottomleft",
  sz.xtab = "small", pearson = TRUE, corplot = TRUE)
}
\arguments{
\item{x, y}{numeric vectors of data values. x and y must have the same length.}

\item{dat}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula.}

\item{nround}{integer indicating the number of decimal places (round).}

\item{xtab}{A logical value indicating whether the output is a xtable}

\item{pos}{a character string indicating the legend location. Options: "bottomright",  "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{pearson}{A logical value indicating whether the text output is Pearson. Default value is TRUE.}

\item{corplot}{A logical value indicating whether the output is a plot. Default value is TRUE.}

\item{A}{character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is "small".}
}
\value{
results:

coeff:

plot
}
\description{
La función quickCor
}
\examples{
data(mtcars)
quickCor(x = "mpg", y = "hp", dat = mtcars,
nround = 3, xtab = FALSE, pearson = TRUE, corplot = TRUE)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{correlation}
\keyword{pearson}
\keyword{plotcor}
\keyword{quickCor}
\keyword{sperman}
