% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCor.R
\name{quickCor}
\alias{quickCor}
\title{quickCor Function}
\usage{
#' @usage \\method{names}{mtc_bis}(x) <- value
}
\arguments{
\item{x, y}{numeric vectors of data values. x and y must have the same length.}

\item{dat}{an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula.}

\item{nround}{integer indicating the number of decimal places (round).}

\item{xtab}{A logical value indicating whether the output is a xtable}

\item{pos}{a character string indicating the legend location. Options: "bottomright",  "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{sz.xtab}{A character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is NULL.}

\item{pearson}{A logical value indicating whether the text output is Pearson. Default value is TRUE.}

\item{corplot}{A logical value indicating whether the output is a plot. Default value is TRUE.}

\item{cex.txt}{character expansion factor. NULL and NA are equivalent to 0.8. This is an absolute measure, not scaled by par("cex") or by setting par("mfrow") or par("mfcol"). Can be a vector.}

\item{cex.main}{settings for main- and sub-title and axis annotation, see \code{\link{title}} and \code{\link{par}}.}

\item{A}{character vector that is inserted just before the tabular environment starts. This can be used to set the font size and a variety of other table settings. Initial backslashes are automatically prefixed, if not supplied by user. Default value is "small".}
}
\value{
results:

coeff:

plot
}
\description{
La función quickCor
}
\examples{
quickCor(x = "mpg", y = "hp", dat = mtc_bis,
nround = 3, xtab = FALSE, pearson = TRUE, corplot = TRUE)
}
\author{
Miriam Mota  \email{miriam.mota@vhir.org}
}
\keyword{correlation}
\keyword{pearson}
\keyword{plotcor}
\keyword{quickCor}
\keyword{sperman}
