% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.quanti.R
\name{summary.quanti}
\alias{summary.quanti}
\title{A summary.quanti Function}
\usage{
\method{summary}{quanti}(x, group = NULL, data, method = "param",
  format = "html", nround = 1, test = NULL, show.pval = TRUE,
  show.all = TRUE, show.n = TRUE, prep2sum = FALSE, sub.ht = TRUE)
}
\arguments{
\item{x}{numeric or integer variable}

\item{group}{factor variable. Outcome}

\item{data}{data frame, list or environment (or object coercible by 'as.data.frame' to a data frame) containing the variables in the model. If they are not found in 'data', the variables are taken from 'environment(formula)'.}

\item{method}{character string indicating the method to test use; possible values are 'param' or 'nonparam'. Default values is 'param'.}

\item{format}{a character string; possible values are ht, r, no. Default value is "ht".}

\item{nround}{integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed (see ‘Details’). Default value is 2.}

\item{test}{character strin indicating the test to use. Possible values are 'anova','t.test','wilcox','kruskal'. Default value is NULL}

\item{show.pval}{logical indicating whether p-value of overall groups significance ('p.overall' column) is displayed or not. Default value is TRUE.}

\item{show.all}{logical indicating whether the 'ALL' column (all data without stratifying by groups) is displayed or not. Default value is FALSE if grouping variable is defined, and FALSE if there are no groups.}

\item{show.n}{ogical indicating whether number of individuals analyzed for each row-variable is displayed or not in the 'descr' table. Default value is TRUE.}

\item{byrow}{logical or NA. Percentage of categorical variables must be reported by rows (TRUE), by columns (FALSE) or by columns and rows to sum up 1 (NA). Default value is FALSE, which means that percentages are reported by columns (withing groups).}
}
\description{
DESCRIPCIO DE LA FUNCIO
}
\examples{
 # set.seed(1)
 # data <- df <- data.frame(MUT = factor(c(rep("A", 12),rep("B",13))),
 #                          var = rnorm(25))

 # summary.quanti(x = "var", data = df)
 # tab <- summary.quanti(x = "var",group = "MUT",data = df)
 # kable(tab$summary,escape = F, row.names = F,align = "c", txt_caption = tab$txt_caption) \%>\%
 #   kable_styling(latex_options = c("striped","hold_position", "repeat_header"), full_width = F, font_size = 14) \%>\%
 #   row_spec(0,background = "#993489", color = "white")
}
\keyword{ci}
\keyword{descriptive}
\keyword{exploratory}
\keyword{qualitative}
\keyword{summary}
